+++ src/libstoriqone-job/script.c
@@ -116,6 +116,6 @@ struct so_value * soj_script_run(struct so_database_connection * db_connect, str
 		free(path);
 	}
 
-	return so_value_pack("{sisosb}", "status", (long int) status, "datas", datas, "should run", should_run);
+	return so_value_pack("{sisosb}", "status", status, "datas", datas, "should run", should_run);
 }
